Pull down the `include_default_values` argument to `to_json`
The `to_json` method does not include the `include_default_values`
option from `to_dict`. While the implementation to `to_json` is
almost litteraly only a call to `to_dict()` (and thus trivial to
replace), it is more convenient for users to not have to reimplement it
when the option is desired.